+++ cmd/vega/node/node.go
@@ -128,23 +128,39 @@ func (n *Command) Run(
 		return err
 	}
 
-	if err := n.startBlockchain(tmHome, network, networkURL); err != nil {
-		return err
-	}
+	// if a chain is being replayed tendermint does this during the initial handshake with the
+	// app and does so synchronously. We to need to set this off in a goroutine so we can catch any
+	// SIGTERM during that replay and shutdown properly
+	errCh := make(chan error)
+	go func() {
+		defer func() {
+			// if a consensus failure happens during replay tendermint panics
+			// we need to catch it so we can call shutdown and then re-panic
+			if r := recover(); r != nil {
+				n.Stop()
+				panic(r)
+			}
+		}()
+		if err := n.startBlockchain(tmHome, network, networkURL); err != nil {
+			errCh <- err
+		}
+		// start the nullblockchain if we are in that mode, it *needs* to be after we've started the gRPC server
+		// otherwise it'll start calling init-chain and all the way before we're ready.
+		if n.conf.Blockchain.ChainProvider == blockchain.ProviderNullChain {
+			n.nullBlockchain.StartServer()
+		}
+	}()
 
 	// at this point all is good, and we should be started, we can
 	// just wait for signals or whatever
 	n.Log.Info("Vega startup complete",
 		logging.String("node-mode", string(n.conf.NodeMode)))
 
-	// start the nullblockchain if we are in that mode, it *needs* to be after we've started the gRPC server
-	// otherwise it'll start calling init-chain and all the way before we're ready.
-	if n.conf.Blockchain.ChainProvider == blockchain.ProviderNullChain {
-		n.nullBlockchain.StartServer()
-	}
-
 	// wait for possible protocol upgrade, or user exist
-	n.wait()
+
+	if err := n.wait(errCh); err != nil {
+		return err
+	}
 
 	// cleanup
 	n.Stop()
@@ -152,20 +168,22 @@ func (n *Command) Run(
 	return nil
 }
 
-func (n *Command) wait() {
+func (n *Command) wait(errCh <-chan error) error {
 	gracefulStop := make(chan os.Signal, 1)
 	signal.Notify(gracefulStop, syscall.SIGTERM, syscall.SIGINT)
-
 	for {
 		select {
 		case version := <-n.protocolUpgrade:
 			n.startProtocolUpgrade(version)
 		case sig := <-gracefulStop:
 			n.Log.Info("Caught signal", logging.String("name", fmt.Sprintf("%+v", sig)))
-			return
+			return nil
+		case e := <-errCh:
+			n.Log.Error("problem starting blockchain", logging.Error(e))
+			return e
 		case <-n.ctx.Done():
 			// nothing to do
-			return
+			return nil
 		}
 	}
 }