+++ core/execution/equity_shares.go
@@ -59,25 +59,7 @@ func (es *EquityShares) OpeningAuctionEnded() {
 		panic("market already left opening auction")
 	}
 	es.openingAuctionEnded = true
-	es.setOpeningAuctionAVG()
 	es.r = num.DecimalZero()
-}
-
-// we just the average entry valuation to the same value
-// for every LP during opening auction.
-func (es *EquityShares) setOpeningAuctionAVG() {
-	// now set average entry valuation for all of them.
-	factor := num.DecimalFromFloat(1.0)
-	if !es.totalVStake.IsZero() && !es.totalPStake.IsZero() {
-		factor = es.totalPStake.Div(es.totalVStake)
-	}
-	for _, v := range es.lps {
-		if v.stake.GreaterThan(v.vStake) {
-			v.vStake = v.stake
-		}
-		v.avg = v.vStake.Mul(factor) // perhaps we ought to move this to a separate loop once the totals have all been updated
-		// v.avg = es.mvp
-	}
 }
 
 func (es *EquityShares) UpdateVStake() {
@@ -96,6 +78,10 @@ func (es *EquityShares) UpdateVStake() {
 		total = total.Add(vStake)
 	}
 	es.totalVStake = total
+}
+
+func (es *EquityShares) GetTotalVStake() num.Decimal {
+	return es.totalVStake
 }
 
 func (es *EquityShares) AvgTradeValue(avg num.Decimal) *EquityShares {
@@ -136,17 +122,10 @@ func (es *EquityShares) SetPartyStake(id string, newStakeU *num.Uint) {
 	}
 	// first time we set the newStake and mvp as avg.
 	if !found {
-		avg := es.mvp
-		es.lps[id] = &lp{
-			stake:  newStake,
-			avg:    avg,
-			vStake: newStake,
-		}
-		es.totalPStake = es.totalPStake.Add(newStake)
-		es.totalVStake = es.totalVStake.Add(newStake)
-		if es.openingAuctionEnded {
-			es.lps[id].avg = es.AvgEntryValuation(id)
-		}
+		// this is technically a delta == new stake -> calculate avg accordingly
+		v = &lp{}
+		es.lps[id] = v
+		es.updateAvgPosDelta(v, newStake, newStake) // delta == new stake
 		return
 	}
 
@@ -161,27 +140,27 @@ func (es *EquityShares) SetPartyStake(id string, newStakeU *num.Uint) {
 		v.stake = newStake
 		es.totalVStake = es.totalVStake.Add(v.vStake)
 		es.totalPStake = es.totalPStake.Add(v.stake)
-		v.avg = v.vStake.Mul(es.totalPStake.Div(es.totalVStake))
 		return
 	}
 
+	es.updateAvgPosDelta(v, delta, newStake)
+}
+
+func (es *EquityShares) updateAvgPosDelta(v *lp, delta, newStake num.Decimal) {
+	// entry valuation == total Virtual stake (before delta is applied)
+	// (average entry valuation) <- (average entry valuation) x S / (S + Delta S) + (entry valuation) x (Delta S) / (S + Delta S)
+	// S being the LP's physical stake, Delta S being the amount by which the stake is increased
 	es.totalVStake = es.totalVStake.Add(delta)
 	es.totalPStake = es.totalPStake.Sub(v.stake).Add(newStake)
-	// stakes were originally assigned _after_ the mustEquity call
-	// average was calculated in the if es.openingAuctionEnded bit
-	// v.avg = v.vStake.Mul(es.totalPStake.Div(es.totalVStake))
-	v.vStake = v.vStake.Add(delta) // increase
+	v.avg = v.avg.Mul(v.vStake).Div(newStake).Add(es.totalVStake.Mul(delta).Div(newStake))
+	// this is the first LP -> no vStake yet
+	v.vStake = v.vStake.Add(delta)
 	v.stake = newStake
-	v.avg = v.vStake.Mul(es.totalPStake.Div(es.totalVStake))
 }
 
 // AvgEntryValuation returns the Average Entry Valuation for a given party.
 func (es *EquityShares) AvgEntryValuation(id string) num.Decimal {
 	if v, ok := es.lps[id]; ok {
-		avg := v.vStake.Mul(es.totalPStake.Div(es.totalVStake))
-		if !avg.Equal(v.avg) {
-			v.avg = avg
-		}
 		return v.avg
 	}
 	return num.DecimalZero()